Be feature/article views

See merge request s10-webmobile1-sub2/S10P12A206!68

+++ back-end/mogul/src/main/java/com/mogul/demo/board/dto/ArticleReadResponse.java
@@ -6,8 +6,8 @@ import lombok.Data;
 
 @Data
 public class ArticleReadResponse extends ArticleResponse{
-	private int userId;
-	private int hit;
+	private Long userId;
+	private Integer hit;
 	private LocalDateTime editedDate;
 
 }
