+++ back-end/mogul/src/main/java/com/mogul/demo/board/repository/ArticleRepository.java
@@ -12,7 +12,7 @@ import com.mogul.demo.board.entity.Article;
 @Repository
 public interface ArticleRepository extends JpaRepository<Article,Long> {
 
-	Optional<Article> findArticleById(int id);
+	Optional<Article> findArticleById(Long id);
 	List<Article> findAllByIsDeletedFalse(Pageable pageable);
 
 }
